configs: Decode "smoke_test" blocks
This is a new experimental kind of block for declaring additional tests
that rely on information that would not naturally be available to
Terraform and therefore require fetching additional data (from data
sources) before checking.

These are also distinguished from our other check types in that smoke
tests are always deferred to the end of the apply phase to check that the
real system is in a valid state, and that failing checks are not treated
as a failing apply step, but rather as a successful apply step with
checks now failing. Separate tooling may then use the check result
information to take some appropriate action if desired, such as raising
an alert.

This commit only implements the configuration decoding. smoke_test blocks
don't actually do anything yet.